+++ app/src/main/java/com/kevinkyang/inventory/ItemChangeListener.java
@@ -1,11 +1,11 @@
 package com.kevinkyang.inventory;
 
 interface ItemChangeListener {
-	void onItemAdded(String name, int quantity, String unit,
+	void onItemAdded(String name, float quantity, String unit,
 					 String type, String expiresDate, String inventory,
 					 boolean inGroceryList);
 
-	void onItemSaved(String name, int quantity, String unit,
+	void onItemSaved(String name, float quantity, String unit,
 					 String type, String expiresDate, String inventory,
 					 boolean inGroceryList, Item item, int position);
